Simple scripts using sounddevice.rec()
These are very simple scripts that collect and plot 2 seconds of data from Digi device

simpleDigi.py collects data and does not apply units. It also collects data from the default sound device, which may or may not be a Digi device.
findDigiImport.py actually looks for an appropriate Digi device. If no Digi device is found, the script exits, otherwise it acquires data and can scales the data with appropriate sensitivity information that is embedded in the name of the Digi device.